`expect` test utility consolidation
