+++ src/test-utilities/expect/expect-loaded/images.ts
