+++ src/test-utilities/expect/expect-loaded/open-street-map-tiles.ts
