+++ src/tutorial/accessibility/interactive.test.ts
@@ -1,5 +1,5 @@
 import { activeElementClassName } from 'test-utilities/active-element-class-name.js'
-import { expectOpenStreetMapTilesLoaded } from 'test-utilities/expect-loaded/open-street-map-tiles.js'
+import { expectOpenStreetMapTilesLoaded } from 'test-utilities/expect/expect-loaded/open-street-map-tiles.js'
 import { pressEnter } from 'test-utilities/input/keypress/enter.js'
 import { pressTab } from 'test-utilities/input/keypress/tab.js'
