+++ src/tutorial/custom-icons/custom-icons.test.ts
@@ -1,5 +1,5 @@
-import { expectImagesLoaded } from 'test-utilities/expect-loaded/images.js'
-import { expectOpenStreetMapTilesLoaded } from 'test-utilities/expect-loaded/open-street-map-tiles.js'
+import { expectImagesLoaded } from 'test-utilities/expect/expect-loaded/images.js'
+import { expectOpenStreetMapTilesLoaded } from 'test-utilities/expect/expect-loaded/open-street-map-tiles.js'
 
 describe('custom icons tutorial', (): void => {
   describe.each([1, 2])(
