+++ src/tutorial/quick-start/quick-start.test.ts
@@ -1,7 +1,7 @@
 import { type BoundingBox } from 'puppeteer'
 
-import { expectImagesLoaded } from 'test-utilities/expect-loaded/images.js'
-import { expectOpenStreetMapTilesLoaded } from 'test-utilities/expect-loaded/open-street-map-tiles.js'
+import { expectImagesLoaded } from 'test-utilities/expect/expect-loaded/images.js'
+import { expectOpenStreetMapTilesLoaded } from 'test-utilities/expect/expect-loaded/open-street-map-tiles.js'
 
 describe('quick start tutorial', (): void => {
   describe.each([1, 2])(
